Model refactoring
- Restructure item and resource type item model classes.
- Removed USE statements.
