+++ app/Item/SimpleItem.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace App\Item;
 
 use App\Interfaces\Item\IModel;
-use App\Models\ItemType\SimpleItem as ItemModel;
+use App\Models\Item\SimpleItem as ItemModel;
 use App\Models\Transformers\Transformer;
 use App\Validators\Fields\ItemType\SimpleItem as ItemTypeSimpleItemValidator;
 use App\Validators\Fields\Validator;
