+++ app/Models/ResourceTypeItem/Summary/AllocatedExpense.php
@@ -1,12 +1,15 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models\ResourceTypeItemType\Summary;
+namespace App\Models\ResourceTypeItem\Summary;
 
 use App\Interfaces\ResourceTypeItem\ISummaryModelCategories;
 use App\Interfaces\ResourceTypeItemType\ISummaryModel;
+use App\Utilities\Model as ModelUtility;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
+use Illuminate\Support\Facades\DB;
 
 /**
  * Item model when fetching data by resource type
@@ -16,11 +19,11 @@ use Illuminate\Database\Query\Builder as QueryBuilder;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategories
+class AllocatedExpense extends Model implements ISummaryModel, ISummaryModelCategories
 {
-    protected $guarded = ['id', 'created_at', 'updated_at'];
+    protected $guarded = ['id', 'actualised_total', 'created_at', 'updated_at'];
     protected $table = 'item';
-    protected $sub_table = 'item_type_simple_expense';
+    protected $sub_table = 'item_type_allocated_expense';
 
     /**
      * Return the summary for all items for the resources in the requested resource type
@@ -35,11 +38,13 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
         array $parameters
     ): array
     {
-        $collection = $this->selectRaw("sum({$this->sub_table}.total) AS total")->
+        $collection = $this->selectRaw("sum({$this->sub_table}.actualised_total) AS total")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join('resource', 'item.resource_id', 'resource.id')->
             join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
             where('resource_type.id', '=', $resource_type_id);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
 
         return $collection->
             get()->
@@ -63,15 +68,153 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
         $collection = $this->selectRaw("
                 resource.id AS id, 
                 resource.name AS `name`, 
-                SUM({$this->sub_table}.total) AS total"
+                SUM({$this->sub_table}.actualised_total) AS total"
             )->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join('resource', 'item.resource_id', 'resource.id')->
             join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
             where('resource_type.id', '=', $resource_type_id);
 
+        $collection = $this->includeUnpublished($collection, $parameters);
+
         return $collection->groupBy('resource.id')->
             orderBy('name')->
+            get()->
+            toArray();
+    }
+
+    /**
+     * Return the summary for all items for the resources in the requested resource
+     * type grouped by year
+     *
+     * @param int $resource_type_id
+     * @param array $parameters
+
+     * @return array
+     */
+    public function yearsSummary(
+        int $resource_type_id,
+        array $parameters
+    ): array
+    {
+        $collection = $this->selectRaw("
+                YEAR({$this->sub_table}.effective_date) as year,
+                SUM({$this->sub_table}.actualised_total) AS total"
+            )->
+            join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
+            join("resource", "resource.id", "item.resource_id")->
+            join("resource_type", "resource_type.id", "resource.resource_type_id")->
+            where("resource_type.id", "=", $resource_type_id);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
+
+        return $collection->groupBy("year")->
+            orderBy("year")->
+            get()->
+            toArray();
+    }
+
+    /**
+     * Return the summary for all items for the resources in the requested resource
+     * type grouped by month for the requested year
+     *
+     * @param integer $resource_type_id
+     * @param integer $year
+     * @param array $parameters
+     *
+     * @return array
+     */
+    public function monthsSummary(
+        int $resource_type_id,
+        int $year,
+        array $parameters
+    ): array
+    {
+        $collection = $this->selectRaw("
+                MONTH({$this->sub_table}.effective_date) as month, 
+                SUM({$this->sub_table}.actualised_total) AS total"
+            )->
+            join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
+            join("resource", "resource.id", "item.resource_id")->
+            join("resource_type", "resource_type.id", "resource.resource_type_id")->
+            where("resource_type.id", "=", $resource_type_id)->
+            where(DB::raw("YEAR({$this->sub_table}.effective_date)"), '=', $year);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
+
+        return $collection->groupBy("month")->
+            orderBy("month")->
+            get()->
+            toArray();
+    }
+
+    /**
+     * Return the summary for all items for the resources in the requested resource
+     * type for a specific year and month
+     *
+     * @param integer $resource_type_id
+     * @param integer $year
+     * @param integer $month
+     * @param array $parameters
+     *
+     * @return array
+     */
+    public function monthSummary(
+        int $resource_type_id,
+        int $year,
+        int $month,
+        array $parameters
+    ): array
+    {
+        $collection = $this->selectRaw("
+                MONTH({$this->sub_table}.effective_date) as month, 
+                SUM({$this->sub_table}.actualised_total) AS total"
+            )->
+            join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
+            join("resource", "resource.id", "item.resource_id")->
+            join("resource_type", "resource_type.id", "resource.resource_type_id")->
+            where("resource_type.id", "=", $resource_type_id)->
+            where(DB::raw("YEAR({$this->sub_table}.effective_date)"), '=', $year)->
+            where(DB::raw("MONTH({$this->sub_table}.effective_date)"), '=', $month);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
+
+        return $collection->groupBy("month")->
+            orderBy("month")->
+            get()->
+            toArray();
+    }
+
+    /**
+     * Return the summary for all items for the resources in the requested resource
+     * type for a specific year
+     *
+     * @param integer $resource_type_id
+     * @param integer $year
+     * @param array $parameters
+     *
+     * @return array
+     */
+    public function yearSummary(
+        int $resource_type_id,
+        int $year,
+        array $parameters
+    ): array
+    {
+        $collection = $this->selectRaw("
+                YEAR({$this->sub_table}.effective_date) as year, 
+                SUM({$this->sub_table}.actualised_total) AS total"
+            )->
+            join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
+            join("resource", "resource.id", "item.resource_id")->
+            join("resource_type", "resource_type.id", "resource.resource_type_id")->
+            where("resource_type.id", "=", $resource_type_id)->
+            where(DB::raw("YEAR({$this->sub_table}.effective_date)"), '=', $year);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
+
+        return $collection->groupBy("year")->
+            orderBy("year")->
             get()->
             toArray();
     }
@@ -94,7 +237,7 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
                 category.id, 
                 category.name AS name, 
                 category.description AS description,
-                SUM({$this->sub_table}.total) AS total")->
+                SUM({$this->sub_table}.actualised_total) AS total")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
@@ -102,6 +245,8 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
             join("category", "category.id", "item_category.category_id")->
             where("category.resource_type_id", "=", $resource_type_id)->
             where("resource_type.id", "=", $resource_type_id);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
 
         return $collection->groupBy("category.id")->
             orderBy("name")->
@@ -129,7 +274,7 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
                 category.id, 
                 category.name AS name, 
                 category.description, 
-                SUM({$this->sub_table}.total) AS total")->
+                SUM({$this->sub_table}.actualised_total) AS total")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
@@ -138,6 +283,8 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
             where("category.resource_type_id", "=", $resource_type_id)->
             where("resource_type.id", "=", $resource_type_id)->
             where("category.id", '=', $category_id);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
 
         return $collection->groupBy("category.id")->
             orderBy("name")->
@@ -153,6 +300,7 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
      * @param int|null $month
      * @param array $parameters
      * @param array $search_parameters
+     * @param array $filter_parameters
      * @return array
      */
     public function filteredSummary(
@@ -162,11 +310,12 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
         int $year = null,
         int $month = null,
         array $parameters = [],
-        array $search_parameters = []
+        array $search_parameters = [],
+        array $filter_parameters = []
     ): array
     {
         $collection = $this->
-            selectRaw("SUM({$this->sub_table}.total) AS total")->
+            selectRaw("SUM({$this->sub_table}.actualised_total) AS total")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
@@ -182,11 +331,26 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
         if ($subcategory_id !== null) {
             $collection->where("sub_category.id", "=", $subcategory_id);
         }
-        if (count($search_parameters) > 0) {
-            foreach ($search_parameters as $field => $search_term) {
-                $collection->where("{$this->sub_table}." . $field, 'LIKE', '%' . $search_term . '%');
-            }
+        if ($year !== null) {
+            $collection->whereRaw(DB::raw("YEAR({$this->sub_table}.effective_date) = {$year}"));
         }
+        if ($month !== null) {
+            $collection->whereRaw(DB::raw("MONTH({$this->sub_table}.effective_date) = {$month}"));
+        }
+
+        $collection = ModelUtility::applySearch(
+            $collection,
+            $this->sub_table,
+            $search_parameters
+        );
+
+        $collection = ModelUtility::applyFiltering(
+            $collection,
+            $this->sub_table,
+            $filter_parameters
+        );
+
+        $collection = $this->includeUnpublished($collection, $parameters);
 
         return $collection->get()->
             toArray();
@@ -212,7 +376,7 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
                 sub_category.id, 
                 sub_category.name AS name, 
                 sub_category.description AS description, 
-                SUM({$this->sub_table}.total) AS total")->
+                SUM({$this->sub_table}.actualised_total) AS total")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
@@ -223,6 +387,8 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
             where("category.resource_type_id", "=", $resource_type_id)->
             where("resource_type.id", "=", $resource_type_id)->
             where("category.id", "=", $category_id);
+
+        $collection = $this->includeUnpublished($collection, $parameters);
 
         return $collection->groupBy("sub_category.id")->
             orderBy("name")->
@@ -252,7 +418,7 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
                 sub_category.id, 
                 sub_category.name AS name, 
                 sub_category.description AS description, 
-                SUM($this->sub_table.total) AS total")->
+                SUM($this->sub_table.actualised_total) AS total")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
@@ -265,9 +431,34 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
             where("category.id", "=", $category_id)->
             where('sub_category.id', '=', $subcategory_id);
 
+        $collection = $this->includeUnpublished($collection, $parameters);
+
         return $collection->groupBy("sub_category.id")->
             orderBy("name")->
             get()->
             toArray();
+    }
+
+    /**
+     * Work out if we should be hiding unpublished items, by default we don't show them
+     *
+     * @param $collection
+     * @param array $parameters
+     *
+     * @return Builder
+     */
+    private function includeUnpublished($collection, array $parameters): Builder
+    {
+        if (
+            array_key_exists('include-unpublished', $parameters) === false ||
+            $parameters['include-unpublished'] === false
+        ) {
+            $collection->where(function ($sql) {
+                $sql->whereNull('item_type_allocated_expense.publish_after')->
+                    orWhereRaw('item_type_allocated_expense.publish_after < NOW()');
+            });
+        }
+
+        return $collection;
     }
 }
