+++ app/ResourceTypeItem/SimpleExpense.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace App\ResourceTypeItem;
 
 use App\Interfaces\ResourceTypeItem\IModel;
-use App\Models\ResourceTypeItemType\SimpleExpense as ItemModel;
+use App\Models\ResourceTypeItem\SimpleExpense as ItemModel;
 use App\Models\Transformers\Transformer;
 
 /**
